+++ x-pack/plugin/ilm/qa/multi-node/src/javaRestTest/java/org/elasticsearch/xpack/ilm/TimeSeriesLifecycleActionsIT.java
@@ -13,7 +13,6 @@ import org.apache.logging.log4j.Logger;
 import org.elasticsearch.client.Request;
 import org.elasticsearch.client.Response;
 import org.elasticsearch.client.ResponseException;
-import org.elasticsearch.client.RestClient;
 import org.elasticsearch.cluster.metadata.IndexMetadata;
 import org.elasticsearch.common.CheckedRunnable;
 import org.elasticsearch.common.Nullable;
@@ -47,7 +46,6 @@ import org.elasticsearch.xpack.core.ilm.SetSingleNodeAllocateStep;
 import org.elasticsearch.xpack.core.ilm.ShrinkAction;
 import org.elasticsearch.xpack.core.ilm.ShrinkStep;
 import org.elasticsearch.xpack.core.ilm.Step;
-import org.elasticsearch.xpack.core.ilm.Step.StepKey;
 import org.elasticsearch.xpack.core.ilm.UpdateRolloverLifecycleDateStep;
 import org.elasticsearch.xpack.core.ilm.WaitForActiveShardsStep;
 import org.elasticsearch.xpack.core.ilm.WaitForRolloverReadyStep;
@@ -70,19 +68,17 @@ import static org.elasticsearch.xpack.TimeSeriesRestDriver.createFullPolicy;
 import static org.elasticsearch.xpack.TimeSeriesRestDriver.createIndexWithSettings;
 import static org.elasticsearch.xpack.TimeSeriesRestDriver.createNewSingletonPolicy;
 import static org.elasticsearch.xpack.TimeSeriesRestDriver.createSnapshotRepo;
-import static org.elasticsearch.xpack.TimeSeriesRestDriver.explain;
 import static org.elasticsearch.xpack.TimeSeriesRestDriver.explainIndex;
 import static org.elasticsearch.xpack.TimeSeriesRestDriver.getNumberOfSegments;
 import static org.elasticsearch.xpack.TimeSeriesRestDriver.getOnlyIndexSettings;
 import static org.elasticsearch.xpack.TimeSeriesRestDriver.getStepKeyForIndex;
+import static org.elasticsearch.xpack.TimeSeriesRestDriver.index;
 import static org.elasticsearch.xpack.TimeSeriesRestDriver.indexDocument;
 import static org.elasticsearch.xpack.TimeSeriesRestDriver.rolloverMaxOneDocCondition;
-import static org.hamcrest.Matchers.allOf;
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.greaterThan;
 import static org.hamcrest.Matchers.greaterThanOrEqualTo;
-import static org.hamcrest.Matchers.hasKey;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.not;
 import static org.hamcrest.Matchers.nullValue;
@@ -90,7 +86,6 @@ import static org.hamcrest.Matchers.nullValue;
 public class TimeSeriesLifecycleActionsIT extends ESRestTestCase {
     private static final Logger logger = LogManager.getLogger(TimeSeriesLifecycleActionsIT.class);
     private static final String FAILED_STEP_RETRY_COUNT_FIELD = "failed_step_retry_count";
-    private static final String IS_AUTO_RETRYABLE_ERROR_FIELD = "is_auto_retryable_error";
 
     private String index;
     private String policy;
@@ -137,84 +132,6 @@ public class TimeSeriesLifecycleActionsIT extends ESRestTestCase {
         assertBusy(() -> assertTrue(indexExists(secondIndex)));
         // asserts that shrink deleted the original index
         assertBusy(() -> assertFalse(indexExists(originalIndex)), 60, TimeUnit.SECONDS);
-        // asserts that the delete phase completed for the managed shrunken index
-        assertBusy(() -> assertFalse(indexExists(shrunkenOriginalIndex)));
-    }
-
-    public void testMoveToAllocateStep() throws Exception {
-        String originalIndex = index + "-000001";
-        createIndexWithSettings(client(), originalIndex, alias, Settings.builder().put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, 4)
-            .put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 0)
-            .put("index.routing.allocation.include._name", "javaRestTest-0")
-            .put(RolloverAction.LIFECYCLE_ROLLOVER_ALIAS, "alias"));
-
-        // create policy
-        createFullPolicy(client(), policy, TimeValue.timeValueHours(10));
-        // update policy on index
-        updatePolicy(originalIndex, policy);
-
-        // move to a step
-        Request moveToStepRequest = new Request("POST", "_ilm/move/" + originalIndex);
-        assertBusy(() -> assertTrue(getStepKeyForIndex(client(), originalIndex).equals(new StepKey("new", "complete", "complete"))));
-        moveToStepRequest.setJsonEntity("{\n" +
-            "  \"current_step\": {\n" +
-            "    \"phase\": \"new\",\n" +
-            "    \"action\": \"complete\",\n" +
-            "    \"name\": \"complete\"\n" +
-            "  },\n" +
-            "  \"next_step\": {\n" +
-            "    \"phase\": \"cold\",\n" +
-            "    \"action\": \"allocate\",\n" +
-            "    \"name\": \"allocate\"\n" +
-            "  }\n" +
-            "}");
-        client().performRequest(moveToStepRequest);
-        assertBusy(() -> assertFalse(indexExists(originalIndex)));
-    }
-
-
-    public void testMoveToRolloverStep() throws Exception {
-        String originalIndex = index + "-000001";
-        String shrunkenOriginalIndex = ShrinkAction.SHRUNKEN_INDEX_PREFIX + originalIndex;
-        String secondIndex = index + "-000002";
-        createIndexWithSettings(client(), originalIndex, alias, Settings.builder().put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, 4)
-            .put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 0)
-            .put("index.routing.allocation.include._name", "javaRestTest-0")
-            .put(RolloverAction.LIFECYCLE_ROLLOVER_ALIAS, alias));
-
-        createFullPolicy(client(), policy, TimeValue.timeValueHours(10));
-        // update policy on index
-        updatePolicy(originalIndex, policy);
-
-        // move to a step
-        Request moveToStepRequest = new Request("POST", "_ilm/move/" + originalIndex);
-        // index document to trigger rollover
-        index(client(), originalIndex, "_id", "foo", "bar");
-        logger.info(getStepKeyForIndex(client(), originalIndex));
-        moveToStepRequest.setJsonEntity("{\n" +
-            "  \"current_step\": {\n" +
-            "    \"phase\": \"new\",\n" +
-            "    \"action\": \"complete\",\n" +
-            "    \"name\": \"complete\"\n" +
-            "  },\n" +
-            "  \"next_step\": {\n" +
-            "    \"phase\": \"hot\",\n" +
-            "    \"action\": \"rollover\",\n" +
-            "    \"name\": \"attempt-rollover\"\n" +
-            "  }\n" +
-            "}");
-        client().performRequest(moveToStepRequest);
-
-        /*
-         * These asserts are in the order that they should be satisfied in, in
-         * order to maximize the time for all operations to complete.
-         * An "out of order" assert here may result in this test occasionally
-         * timing out and failing inappropriately.
-         */
-        // asserts that rollover was called
-        assertBusy(() -> assertTrue(indexExists(secondIndex)));
-        // asserts that shrink deleted the original index
-        assertBusy(() -> assertFalse(indexExists(originalIndex)), 30, TimeUnit.SECONDS);
         // asserts that the delete phase completed for the managed shrunken index
         assertBusy(() -> assertFalse(indexExists(shrunkenOriginalIndex)));
     }
@@ -888,16 +805,16 @@ public class TimeSeriesLifecycleActionsIT extends ESRestTestCase {
         policy = otherPolicy;
         createNewSingletonPolicy(client(), policy, "delete", new DeleteAction(), TimeValue.timeValueHours(13));
 
-        createIndexWithSettingsNoAlias(managedIndex1, Settings.builder()
-            .put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, randomIntBetween(1,10))
+        createIndexWithSettings(client(), managedIndex1, Settings.builder()
+            .put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, randomIntBetween(1, 10))
             .put(LifecycleSettings.LIFECYCLE_NAME_SETTING.getKey(), originalPolicy));
-        createIndexWithSettingsNoAlias(managedIndex2, Settings.builder()
-            .put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, randomIntBetween(1,10))
+        createIndexWithSettings(client(), managedIndex2, Settings.builder()
+            .put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, randomIntBetween(1, 10))
             .put(LifecycleSettings.LIFECYCLE_NAME_SETTING.getKey(), originalPolicy));
-        createIndexWithSettingsNoAlias(unmanagedIndex, Settings.builder()
-            .put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, randomIntBetween(1,10)));
-        createIndexWithSettingsNoAlias(managedByOtherPolicyIndex, Settings.builder()
-            .put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, randomIntBetween(1,10))
+        createIndexWithSettings(client(), unmanagedIndex, Settings.builder()
+            .put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, randomIntBetween(1, 10)));
+        createIndexWithSettings(client(), managedByOtherPolicyIndex, Settings.builder()
+            .put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, randomIntBetween(1, 10))
             .put(LifecycleSettings.LIFECYCLE_NAME_SETTING.getKey(), otherPolicy));
 
         Request deleteRequest = new Request("DELETE", "_ilm/policy/" + originalPolicy);
@@ -951,80 +868,6 @@ public class TimeSeriesLifecycleActionsIT extends ESRestTestCase {
         assertBusy(() -> assertThat(getStepKeyForIndex(client(), originalIndex), equalTo(PhaseCompleteStep.finalStep("hot").getKey())));
     }
 
-    public void testMoveToInjectedStep() throws Exception {
-        String shrunkenIndex = ShrinkAction.SHRUNKEN_INDEX_PREFIX + index;
-        createNewSingletonPolicy(client(), policy, "warm", new ShrinkAction(1), TimeValue.timeValueHours(12));
-
-        createIndexWithSettings(client(), index, alias, Settings.builder().put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, 3)
-            .put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 0)
-            .put(LifecycleSettings.LIFECYCLE_NAME, policy)
-            .put(RolloverAction.LIFECYCLE_ROLLOVER_ALIAS, alias));
-
-        assertBusy(() -> assertThat(getStepKeyForIndex(client(), index), equalTo(new StepKey("new", "complete", "complete"))));
-
-        // Move to a step from the injected unfollow action
-        Request moveToStepRequest = new Request("POST", "_ilm/move/" + index);
-        moveToStepRequest.setJsonEntity("{\n" +
-            "  \"current_step\": { \n" +
-            "    \"phase\": \"new\",\n" +
-            "    \"action\": \"complete\",\n" +
-            "    \"name\": \"complete\"\n" +
-            "  },\n" +
-            "  \"next_step\": { \n" +
-            "    \"phase\": \"warm\",\n" +
-            "    \"action\": \"unfollow\",\n" +
-            "    \"name\": \"wait-for-indexing-complete\"\n" +
-            "  }\n" +
-            "}");
-        // If we get an OK on this request we have successfully moved to the injected step
-        assertOK(client().performRequest(moveToStepRequest));
-
-        // Make sure we actually move on to and execute the shrink action
-        assertBusy(() -> {
-            assertTrue(indexExists(shrunkenIndex));
-            assertTrue(aliasExists(shrunkenIndex, index));
-            assertThat(getStepKeyForIndex(client(), shrunkenIndex), equalTo(PhaseCompleteStep.finalStep("warm").getKey()));
-        }, 30, TimeUnit.SECONDS);
-    }
-
-    @AwaitsFix(bugUrl = "https://github.com/elastic/elasticsearch/issues/53612")
-    public void testMoveToStepRereadsPolicy() throws Exception {
-        createNewSingletonPolicy(client(), policy, "hot", new RolloverAction(null, TimeValue.timeValueHours(1), null), TimeValue.ZERO);
-
-        createIndexWithSettings(client(), "test-1", alias, Settings.builder()
-                .put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, 1)
-                .put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 0)
-                .put(LifecycleSettings.LIFECYCLE_NAME, policy)
-                .put(RolloverAction.LIFECYCLE_ROLLOVER_ALIAS, alias),
-            true);
-
-        assertBusy(() -> assertThat(getStepKeyForIndex(client(), "test-1"),
-            equalTo(new StepKey("hot", "rollover", "check-rollover-ready"))));
-
-        createNewSingletonPolicy(client(), policy, "hot", new RolloverAction(null, TimeValue.timeValueSeconds(1), null), TimeValue.ZERO);
-
-        // Move to the same step, which should re-read the policy
-        Request moveToStepRequest = new Request("POST", "_ilm/move/test-1");
-        moveToStepRequest.setJsonEntity("{\n" +
-            "  \"current_step\": { \n" +
-            "    \"phase\": \"hot\",\n" +
-            "    \"action\": \"rollover\",\n" +
-            "    \"name\": \"check-rollover-ready\"\n" +
-            "  },\n" +
-            "  \"next_step\": { \n" +
-            "    \"phase\": \"hot\",\n" +
-            "    \"action\": \"rollover\",\n" +
-            "    \"name\": \"check-rollover-ready\"\n" +
-            "  }\n" +
-            "}");
-        assertOK(client().performRequest(moveToStepRequest));
-
-        // Make sure we actually rolled over
-        assertBusy(() -> {
-            indexExists("test-000002");
-        });
-    }
-
     public void testCanStopILMWithPolicyUsingNonexistentPolicy() throws Exception {
         createIndexWithSettings(client(), index, alias, Settings.builder().put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, 1)
             .put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 0)
@@ -1045,80 +888,6 @@ public class TimeSeriesLifecycleActionsIT extends ESRestTestCase {
         // Re-start ILM so that subsequent tests don't fail
         Request startILMReqest = new Request("POST", "_ilm/start");
         assertOK(client().performRequest(startILMReqest));
-    }
-
-    public void testExplainFilters() throws Exception {
-        String goodIndex = index + "-good-000001";
-        String errorIndex = index + "-error";
-        String nonexistantPolicyIndex = index + "-nonexistant-policy";
-        String unmanagedIndex = index + "-unmanaged";
-
-        createFullPolicy(client(), policy, TimeValue.ZERO);
-
-        {
-            // Create a "shrink-only-policy"
-            Map<String, LifecycleAction> warmActions = new HashMap<>();
-            warmActions.put(ShrinkAction.NAME, new ShrinkAction(17));
-            Map<String, Phase> phases = new HashMap<>();
-            phases.put("warm", new Phase("warm", TimeValue.ZERO, warmActions));
-            LifecyclePolicy lifecyclePolicy = new LifecyclePolicy("shrink-only-policy", phases);
-            // PUT policy
-            XContentBuilder builder = jsonBuilder();
-            lifecyclePolicy.toXContent(builder, null);
-            final StringEntity entity = new StringEntity(
-                "{ \"policy\":" + Strings.toString(builder) + "}", ContentType.APPLICATION_JSON);
-            Request request = new Request("PUT", "_ilm/policy/shrink-only-policy");
-            request.setEntity(entity);
-            assertOK(client().performRequest(request));
-        }
-
-        createIndexWithSettings(client(), goodIndex, alias, Settings.builder()
-            .put(RolloverAction.LIFECYCLE_ROLLOVER_ALIAS, alias)
-            .put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 0)
-            .put(LifecycleSettings.LIFECYCLE_NAME, policy)
-        );
-        createIndexWithSettingsNoAlias(errorIndex, Settings.builder()
-            .put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 0)
-            .put(LifecycleSettings.LIFECYCLE_NAME, "shrink-only-policy")
-        );
-        createIndexWithSettingsNoAlias(nonexistantPolicyIndex, Settings.builder()
-            .put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 0)
-            .put(LifecycleSettings.LIFECYCLE_NAME, randomValueOtherThan(policy, () -> randomAlphaOfLengthBetween(3,10))));
-        createIndexWithSettingsNoAlias(unmanagedIndex, Settings.builder()
-            .put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 0));
-
-        assertBusy(() -> {
-            Map<String, Map<String, Object>> explainResponse = explain(client(), index + "*", false, false);
-            assertNotNull(explainResponse);
-            assertThat(explainResponse,
-                allOf(hasKey(goodIndex), hasKey(errorIndex), hasKey(nonexistantPolicyIndex), hasKey(unmanagedIndex)));
-
-            Map<String, Map<String, Object>> onlyManagedResponse = explain(client(), index + "*", false, true);
-            assertNotNull(onlyManagedResponse);
-            assertThat(onlyManagedResponse, allOf(hasKey(goodIndex), hasKey(errorIndex), hasKey(nonexistantPolicyIndex)));
-            assertThat(onlyManagedResponse, not(hasKey(unmanagedIndex)));
-
-            Map<String, Map<String, Object>> onlyErrorsResponse = explain(client(), index + "*", true, true);
-            assertNotNull(onlyErrorsResponse);
-            assertThat(onlyErrorsResponse, allOf(hasKey(errorIndex), hasKey(nonexistantPolicyIndex)));
-            assertThat(onlyErrorsResponse, allOf(not(hasKey(goodIndex)), not(hasKey(unmanagedIndex))));
-        });
-    }
-
-    public void testExplainIndexContainsAutomaticRetriesInformation() throws Exception {
-        createFullPolicy(client(), policy, TimeValue.ZERO);
-
-        // create index without alias so the rollover action fails and is retried
-        createIndexWithSettingsNoAlias(index, Settings.builder()
-            .put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 0)
-            .put(LifecycleSettings.LIFECYCLE_NAME, policy)
-        );
-
-        assertBusy(() -> {
-            Map<String, Object> explainIndex = explainIndex(client(), index);
-            assertThat((Integer) explainIndex.get(FAILED_STEP_RETRY_COUNT_FIELD), greaterThanOrEqualTo(1));
-            assertThat(explainIndex.get(IS_AUTO_RETRYABLE_ERROR_FIELD), is(true));
-        });
     }
 
     public void testILMRolloverRetriesOnReadOnlyBlock() throws Exception {
@@ -1742,26 +1511,6 @@ public class TimeSeriesLifecycleActionsIT extends ESRestTestCase {
         assertEquals("hot", stepKey.getPhase());
         assertEquals(RolloverAction.NAME, stepKey.getAction());
         assertEquals(WaitForRolloverReadyStep.NAME, stepKey.getName());
-    }
-
-    private void createIndexWithSettingsNoAlias(String index, Settings.Builder settings) throws IOException {
-        Request request = new Request("PUT", "/" + index);
-        request.setJsonEntity("{\n \"settings\": " + Strings.toString(settings.build())
-            + "}");
-        client().performRequest(request);
-        // wait for the shards to initialize
-        ensureGreen(index);
-    }
-
-    private static void index(RestClient client, String index, String id, Object... fields) throws IOException {
-        XContentBuilder document = jsonBuilder().startObject();
-        for (int i = 0; i < fields.length; i += 2) {
-            document.field((String) fields[i], fields[i + 1]);
-        }
-        document.endObject();
-        final Request request = new Request("POST", "/" + index + "/_doc/" + id);
-        request.setJsonEntity(Strings.toString(document));
-        assertOK(client.performRequest(request));
     }
 
     @Nullable
