+++ x-pack/plugin/ilm/qa/rest/src/yamlRestTest/resources/rest-api-spec/test/ilm/40_explain_lifecycle.yml
@@ -101,132 +101,6 @@ teardown:
         policy: "my_moveable_timeseries_lifecycle"
 
 ---
-"Test Basic Lifecycle Explain":
-
-  - do:
-      ilm.explain_lifecycle:
-        index: "my_index"
-
-  - is_true: indices.my_index.managed
-  - match: { indices.my_index.index: "my_index" }
-  - match: { indices.my_index.policy: "my_moveable_timeseries_lifecycle" }
-  - match: { indices.my_index.phase: "new" }
-  - match: { indices.my_index.action: "complete" }
-  - match: { indices.my_index.step: "complete" }
-  - is_true: indices.my_index.phase_time_millis
-  - is_true: indices.my_index.age
-  - is_true: indices.my_index.phase_execution
-  - is_false: indices.my_index.failed_step
-  - is_false: indices.my_index.step_info
-
-  - is_false: indices.my_index2
-  - is_false: indices.another_index
-  - is_false: indices.unmanaged_index
-  - is_false: indices.index_with_policy_that_doesnt_exist
-
----
-"Test Wildcard Index Lifecycle Explain":
-
-  - do:
-      ilm.explain_lifecycle:
-        index: "my_*"
-
-  - is_true: indices.my_index.managed
-  - match: { indices.my_index.index: "my_index" }
-  - match: { indices.my_index.policy: "my_moveable_timeseries_lifecycle" }
-  - match: { indices.my_index.phase: "new" }
-  - match: { indices.my_index.action: "complete" }
-  - match: { indices.my_index.step: "complete" }
-  - is_true: indices.my_index.phase_time_millis
-  - is_true: indices.my_index.age
-  - is_true: indices.my_index.phase_execution
-  - is_false: indices.my_index.failed_step
-  - is_false: indices.my_index.step_info
-
-  - is_true: indices.my_index2.managed
-  - match: { indices.my_index2.index: "my_index2" }
-  - match: { indices.my_index2.policy: "my_moveable_timeseries_lifecycle" }
-  - match: { indices.my_index2.phase: "new" }
-  - match: { indices.my_index2.action: "complete" }
-  - match: { indices.my_index2.step: "complete" }
-  - is_true: indices.my_index2.phase_time_millis
-  - is_true: indices.my_index2.age
-  - is_true: indices.my_index2.phase_execution
-  - is_false: indices.my_index2.failed_step
-  - is_false: indices.my_index2.step_info
-
-  - is_false: indices.another_index
-  - is_false: indices.unmanaged_index
-  - is_false: indices.index_with_policy_that_doesnt_exist
-
-
----
-"Test All Indexes Lifecycle Explain":
-  - skip:
-      reason: https://github.com/elastic/elasticsearch/issues/47275
-      version: "6.7.0 - "
-
-  - do:
-      ilm.explain_lifecycle:
-        index: "*"
-
-  - is_true: indices.my_index.managed
-  - match: { indices.my_index.index: "my_index" }
-  - match: { indices.my_index.policy: "my_moveable_timeseries_lifecycle" }
-  - match: { indices.my_index.phase: "new" }
-  - match: { indices.my_index.action: "complete" }
-  - match: { indices.my_index.step: "complete" }
-  - is_true: indices.my_index.phase_time_millis
-  - is_true: indices.my_index.age
-  - is_true: indices.my_index.phase_execution
-  - is_false: indices.my_index.failed_step
-  - is_false: indices.my_index.step_info
-
-  - is_true: indices.my_index2.managed
-  - match: { indices.my_index2.index: "my_index2" }
-  - match: { indices.my_index2.policy: "my_moveable_timeseries_lifecycle" }
-  - match: { indices.my_index2.phase: "new" }
-  - match: { indices.my_index2.action: "complete" }
-  - match: { indices.my_index2.step: "complete" }
-  - is_true: indices.my_index2.phase_time_millis
-  - is_true: indices.my_index2.age
-  - is_true: indices.my_index2.phase_execution
-  - is_false: indices.my_index2.failed_step
-  - is_false: indices.my_index2.step_info
-
-  - is_true: indices.another_index.managed
-  - match: { indices.another_index.index: "another_index" }
-  - match: { indices.another_index.policy: "my_moveable_timeseries_lifecycle" }
-  - match: { indices.another_index.phase: "new" }
-  - match: { indices.another_index.action: "complete" }
-  - match: { indices.another_index.step: "complete" }
-  - is_true: indices.another_index.phase_time_millis
-  - is_true: indices.another_index.age
-  - is_true: indices.another_index.phase_execution
-  - is_false: indices.another_index.failed_step
-  - is_false: indices.another_index.step_info
-
-  - match: { indices.unmanaged_index.index: "unmanaged_index" }
-  - is_false: indices.unmanaged_index.managed
-  - is_false: indices.unmanaged_index.policy
-  - is_false: indices.unmanaged_index.phase
-  - is_false: indices.unmanaged_index.action
-  - is_false: indices.unmanaged_index.step
-  - is_false: indices.unmanaged.age
-  - is_false: indices.another_index.failed_step
-  - is_false: indices.another_index.step_info
-
-  - match: { indices.index_with_policy_that_doesnt_exist.index: "index_with_policy_that_doesnt_exist" }
-  - match: { indices.index_with_policy_that_doesnt_exist.policy: "a_policy_that_doesnt_exist" }
-  - match: { indices.index_with_policy_that_doesnt_exist.step_info.reason: "policy [a_policy_that_doesnt_exist] does not exist" }
-  - is_true: indices.index_with_policy_that_doesnt_exist.managed
-  - is_false: indices.index_with_policy_that_doesnt_exist.phase
-  - is_false: indices.index_with_policy_that_doesnt_exist.action
-  - is_false: indices.index_with_policy_that_doesnt_exist.step
-  - is_false: indices.index_with_policy_that_doesnt_exist.age
-  - is_false: indices.index_with_policy_that_doesnt_exist.failed_step
-
----
 "Test Unmanaged Index Lifecycle Explain":
 
   - do:
